+++ packages/coding-agent/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed the status-line session segment showing identical IDs for sessions created close together by displaying a 13-character prefix by default and supporting `statusLine.segmentOptions.session.length` overrides ([#9572](https://github.com/can1357/oh-my-pi/issues/9572)).
+
 ## [18.0.4] - 2026-08-24
 
 ### Added